requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-ollama
chromadb
langchain_huggingface
sentence_transformers
faiss_cpu
langchain_chroma
streamlit
langchain_groq
fastapi
langserve
